celcius to faranheit

+++ assign1.py
@@ -1,5 +1,5 @@
 C=input("Enter celcius temperature: ")
-C=int(C)
+C=float(C)
 F=((9*C)/5)+32
 
 print("Faranheit Temperature is " + str(F))
